requirements.txt
python-telegram-bot
requests
aiohttp
matplotlib
mplfinance